third_party/blink/web_tests/platform/generic/external/wpt/xhr/no-utf16-json-expected.txt

This is a testharness.js-based test.
FAIL JSON asa UTF-16 with BOM should decode as UTF-8 (and, therefore, fail) assert_equals: expected null but got object "[object Object]"
Harness: the test ran to completion.
